+++ _randomgen/doc/source/change-log.rst
@@ -4,6 +4,8 @@ Change Log
 Changes since v1.14
 ===================
 
+- Synced upstream changes in permutation
+- Synced upstream doc fixes
 - Added absolute_import to avoid import noise on Python 2.7
 - Add legacy generator which allows NumPy replication
 - Improve type handling of integers